requirements.txt
Flask==2.1.2
Werkzeug==2.0.3
Pillow==9.3.0
tqdm==4.66.1
regex==2023.12.25
fsspec==2023.12.2
opencv-python
flask_cors
imageio
pillow-avif-plugin
pillow-heif
websocket-client
bs4
boto3
python-dotenv
flask-socketio
flask-bcrypt
pymongo
PyJWT
pymongo[srv]
segment_anything
gunicorn
pydantic